# backend/docs/swagger.yaml
basePath: /api/
definitions:
...
  models.Host:
    properties:
      date:
        type: string
      deviceType:
        type: string
      dns:
        type: string
      hw:
        type: string
      id:
        type: integer
      iface:
        type: string
      ip:
        type: string
      known:
        type: integer
      location:
        type: string
      mac:
        type: string
      name:
        type: string
      notes:
        type: string
      now:
        type: integer
      owner:
        type: string
      pinned:
        type: boolean
      tags:
        items:
          type: string
        type: array
    type: object
...
info:
  contact:
    url: https://github.com/godlev/LANnventory
  description: Local network inventory and monitoring application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: LANnventory API
  version: 0.1.0-beta.2
paths:
...
  /edit/{id}/{name}/{known}:
    get:
      description: Update a host's name and optionally toggle its "known" status
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: string
      - description: New name for the host
        in: path
        name: name
        required: true
        type: string
      - description: Pass 'toggle' to flip the known/unknown status
        in: path
        name: known
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Edit host
      tags:
      - hosts
...
  /host/{id}:
    get:
      description: Retrieve detailed information about a host by its unique ID
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Host'
      summary: Get host by ID
      tags:
      - hosts
...
  /host/{id}/type:
    patch:
      consumes:
      - application/json
      description: Update only a host's manually assigned device type
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: string
      - description: Device type payload
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Host'
      summary: Set host device type
      tags:
      - hosts
  /host/add/{mac}:
    get:
      description: |-
        Add host by MAC, with optional Name, IP, Hardware
        Returns `models.Host` with this MAC form DB, either just added or existing
      parameters:
      - description: Host MAC
        in: path
        name: mac
        required: true
        type: string
      - description: Name
        in: query
        name: name
        type: string
      - description: IP
        in: query
        name: ip
        type: string
      - description: Hardware
        in: query
        name: hw
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Host'
      summary: Add host manually
      tags:
      - hosts
  /host/del/{id}:
    get:
      description: Remove a host from the database by its unique ID
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete host
      tags:
      - hosts
...
swagger: "2.0"